% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_fit.R
\name{correct_fit}
\alias{correct_fit}
\title{Fit a time series model with correct data size}
\usage{
correct_fit(y, p, k_max)
}
\arguments{
\item{y}{data}

\item{p}{number of lags}

\item{k_max}{max lag length within the set of models}
}
\description{
Fit a time series model with correct data size
}
