% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_hv.R
\name{cv_hv}
\alias{cv_hv}
\title{criterion function for hv cross validation in Racine (2000), JoE}
\usage{
cv_hv(y, p, h, v)
}
\arguments{
\item{y}{data}

\item{p}{number of lags}

\item{v}{size of validation set = 2v+1}
}
\value{
value of criterion function
}
\description{
criterion function for hv cross validation in Racine (2000), JoE
}
