% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{dgp1}
\alias{dgp1}
\title{Data Generating Process: AR(1) with std normal error}
\usage{
dgp1(sample_size, lag1, y0)
}
\arguments{
\item{sample_size}{sample size}

\item{lag1}{AR coefficient}

\item{y0}{initial value of y}
}
\value{
y of length sample_size
}
\description{
Data Generating Process: AR(1) with std normal error
}
