% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP2.R
\name{dgp2}
\alias{dgp2}
\title{Data Generating Process: MA(1) with std normal error}
\usage{
dgp2(sample_size, lag1, e0)
}
\arguments{
\item{sample_size}{sample size}

\item{lag1}{MA coefficient}

\item{e0}{initial value of epsilon}
}
\value{
y of length sample_size
}
\description{
Data Generating Process: MA(1) with std normal error
}
